fix: Add task content validation to prevent empty content processing (#2616)
